rework generation
