+++ tools/serverpod_cli/bin/serverpod_cli.dart
@@ -5,7 +5,7 @@ import 'package:args/args.dart';
 import 'package:colorize/colorize.dart';
 
 import 'package:serverpod_cli/src/analytics/analytics.dart';
-import 'package:serverpod_cli/src/analyzer/dart/protocol_analyzer.dart';
+import 'package:serverpod_cli/src/analyzer/dart/endpoints_analyzer.dart';
 import 'package:serverpod_cli/src/create/create.dart';
 import 'package:serverpod_cli/src/downloads/resource_manager.dart';
 import 'package:serverpod_cli/src/generated/version.dart';
@@ -205,7 +205,8 @@ Future<void> _main(List<String> args) async {
         return;
       }
 
-      var analyzer = ProtocolDartFileAnalyzer(config.endpointsSourcePath);
+      var analyzer =
+          EndpointsAnalyzer(Directory(config.relativeEndpointsSourcePath));
       await performGenerate(
         verbose: verbose,
         config: config,
