+++ tools/serverpod_cli/lib/src/analyzer/code_analysis_collector.dart
@@ -1,0 +1,11 @@
+import 'package:source_span/source_span.dart';
+
+abstract class CodeAnalysisCollector {
+  void addError(SourceSpanException error);
+
+  void addErrors(List<SourceSpanException> errors);
+
+  void printErrors();
+
+  void clearErrors();
+}
