+++ tools/serverpod_cli/lib/src/analyzer/protocol_definition.dart
@@ -1,0 +1,18 @@
+import 'dart/definitions.dart';
+import 'yaml/definitions.dart';
+
+/// Defines a projects protocol.
+/// This does not include stuff the [ProtocolYamlFileAnalyzer] analyzed.
+class ProtocolDefinition {
+  /// The endpoints that are a part of this protocol.
+  /// This does not include endpoints from other modules or package:serverpod.
+  final List<EndpointDefinition> endpoints;
+
+  final List<ProtocolEntityDefinition> entities;
+
+  /// Create a new [ProtocolDefinition].
+  const ProtocolDefinition({
+    required this.endpoints,
+    required this.entities,
+  });
+}
