+++ tools/serverpod_cli/lib/src/generator/generator_continuous.dart
@@ -1,5 +1,5 @@
 import 'package:async/async.dart';
-import 'package:serverpod_cli/src/analyzer/dart/protocol_analyzer.dart';
+import 'package:serverpod_cli/src/analyzer/dart/endpoints_analyzer.dart';
 import 'package:watcher/watcher.dart';
 
 import 'config.dart';
@@ -8,12 +8,12 @@ import 'generator.dart';
 void performGenerateContinuously(
   bool verbose,
   GeneratorConfig config, {
-  required ProtocolDartFileAnalyzer analyzer,
+  required EndpointsAnalyzer analyzer,
 }) async {
   if (verbose) print('Starting up continuous generator');
 
-  var watcherClasses = DirectoryWatcher(config.protocolSourcePath);
-  var watcherEndpoints = DirectoryWatcher(config.endpointsSourcePath);
+  var watcherClasses = DirectoryWatcher(config.relativeProtocolSourcePath);
+  var watcherEndpoints = DirectoryWatcher(config.relativeEndpointsSourcePath);
 
   await for (WatchEvent event
       in StreamGroup.merge([watcherClasses.events, watcherEndpoints.events])) {
