+++ tools/serverpod_cli/lib/src/run/runner.dart
@@ -1,279 +1,282 @@
-import 'dart:async';
-import 'dart:io';
+// import 'dart:async';
+// import 'dart:io';
 
-import '../config_info/config_info.dart';
-import '../create/port_checker.dart';
-import '../generator/config.dart';
-import '../generator/generator.dart';
-import '../analyzer/dart/protocol_analyzer.dart';
-import '../generator/protocol_generator.dart';
-import '../generator/code_analysis_collector.dart';
-import '../port_scanner/port_scanner.dart';
-import '../util/command_line_tools.dart';
-import '../util/print.dart';
-import '../util/process_killer_extension.dart';
-import 'file_watcher.dart';
+// import 'package:serverpod_cli/src/generator/code_generation_collector.dart';
 
-void performRun(bool verbose) async {
-  // TODO: Fix Docker management
-  bool runDocker = false;
+// import '../config_info/config_info.dart';
+// import '../create/port_checker.dart';
+// import '../generator/config.dart';
+// import '../generator/generator.dart';
+// import '../analyzer/dart/endpoints_analyzer.dart';
+// import '../analyzer/code_analysis_collector.dart';
+// import '../port_scanner/port_scanner.dart';
+// import '../util/command_line_tools.dart';
+// import '../util/print.dart';
+// import '../util/process_killer_extension.dart';
+// import 'file_watcher.dart';
 
-  var config = GeneratorConfig.load();
-  if (config == null) {
-    return;
-  }
+// void performRun(bool verbose) async {
+//   // TODO: Fix Docker management
+//   bool runDocker = false;
 
-  var configInfo = ConfigInfo('development');
+//   var config = GeneratorConfig.load();
+//   if (config == null) {
+//     return;
+//   }
 
-  // Check if all required ports are available and Docker is running.
-  var dockerConfigured = (await CommandLineTools.existsCommand('docker') &&
-          await CommandLineTools.isDockerRunning()) ||
-      !runDocker;
+//   var configInfo = ConfigInfo('development');
 
-  var mainPortAvailable =
-      await isNetworkPortAvailable(configInfo.config.apiServer.port);
-  var servicePortAvailable =
-      await isNetworkPortAvailable(configInfo.config.insightsServer.port);
+//   // Check if all required ports are available and Docker is running.
+//   var dockerConfigured = (await CommandLineTools.existsCommand('docker') &&
+//           await CommandLineTools.isDockerRunning()) ||
+//       !runDocker;
 
-  if (!dockerConfigured || !mainPortAvailable || !servicePortAvailable) {
-    var strIssue =
-        'There are some issues with your setup that will prevent your Serverpod project from launching.';
-    // var strIssueDocker =
-    //     'You do not have Docker installed or it is not running. Serverpod uses Docker to run Postgres and Redis. It\'s recommended that you install Docker Desktop from https://www.docker.com/get-started but you can also install and configure Postgres and Redis manually and run this command with the --no-run-docker flag added.';
-    var strIssueMainPort =
-        'The public api port (${configInfo.config.apiServer.port}) is occupied by another application or you are running another instance of Serverpod.';
-    var strIssueServicePort =
-        'The service insights port (${configInfo.config.insightsServer.port}) is occupied by another application or you are running another instance of Serverpod.';
+//   var mainPortAvailable =
+//       await isNetworkPortAvailable(configInfo.config.apiServer.port);
+//   var servicePortAvailable =
+//       await isNetworkPortAvailable(configInfo.config.insightsServer.port);
 
-    printww(strIssue);
+//   if (!dockerConfigured || !mainPortAvailable || !servicePortAvailable) {
+//     var strIssue =
+//         'There are some issues with your setup that will prevent your Serverpod project from launching.';
+//     // var strIssueDocker =
+//     //     'You do not have Docker installed or it is not running. Serverpod uses Docker to run Postgres and Redis. It\'s recommended that you install Docker Desktop from https://www.docker.com/get-started but you can also install and configure Postgres and Redis manually and run this command with the --no-run-docker flag added.';
+//     var strIssueMainPort =
+//         'The public api port (${configInfo.config.apiServer.port}) is occupied by another application or you are running another instance of Serverpod.';
+//     var strIssueServicePort =
+//         'The service insights port (${configInfo.config.insightsServer.port}) is occupied by another application or you are running another instance of Serverpod.';
 
-    // if (!dockerConfigured) {
-    //   printww('');
-    //   printww(strIssueDocker);
-    // }
-    if (!mainPortAvailable) {
-      printww('');
-      printww(strIssueMainPort);
-    }
-    if (!servicePortAvailable) {
-      printww('');
-      printww(strIssueServicePort);
-    }
-    return;
-  }
+//     printww(strIssue);
 
-  // It looks like we are ready to go.
-  printwwln('Starting Serverpod.');
-  printww(' • Automatic generate and reload are enabled.');
-  // if (runDocker) {
-  //   print(' • Running Postgres and Redis in Docker container.');
-  // }
-  printww('');
+//     // if (!dockerConfigured) {
+//     //   printww('');
+//     //   printww(strIssueDocker);
+//     // }
+//     if (!mainPortAvailable) {
+//       printww('');
+//       printww(strIssueMainPort);
+//     }
+//     if (!servicePortAvailable) {
+//       printww('');
+//       printww(strIssueServicePort);
+//     }
+//     return;
+//   }
 
-  // Do an initial serverpod generate.
-  printww('Spinning up serverpod generate (this can take a few seconds).');
-  // TODO: Fix!
-  // performGenerateClasses(verbose);
-  var collector = CodeAnalysisCollector();
+//   // It looks like we are ready to go.
+//   printwwln('Starting Serverpod.');
+//   printww(' • Automatic generate and reload are enabled.');
+//   // if (runDocker) {
+//   //   print(' • Running Postgres and Redis in Docker container.');
+//   // }
+//   printww('');
 
-  var analyzer = ProtocolDartFileAnalyzer(config.endpointsSourcePath);
+//   // Do an initial serverpod generate.
+//   printww('Spinning up serverpod generate (this can take a few seconds).');
+//   // TODO: Fix!
+//   // performGenerateClasses(verbose);
+//   var collector = CodeGenerationCollector();
 
-  var protocolDefinition = await analyzer.analyze(
-    verbose: verbose,
-    collector: collector,
-  );
-  collector.printErrors();
-  await performGenerateProtocol(
-    verbose: verbose,
-    protocolDefinition: protocolDefinition,
-    collector: collector,
-    codeGenerator: (spec) => generateCode(spec, true),
-    config: config,
-  );
+//   var analyzer =
+//       EndpointsAnalyzer(Directory(config.relativeEndpointsSourcePath));
 
-  // Analyze the code
-  var errors = await analyzer.getErrors();
-  if (errors.isNotEmpty) {
-    for (var error in errors) {
-      print(error);
-    }
-    return;
-  }
+//   var protocolDefinition = await analyzer.analyze(
+//     verbose: verbose,
+//     collector: collector,
+//   );
+//   collector.printErrors();
 
-  // Generate continuously and hot reload.
-  var serverRunner = _ServerRunner();
-  // var dockerRunner = _DockerRunner();
-  var generatingAndReloading = false;
+//   await performGenerateProtocol(
+//     verbose: verbose,
+//     protocolDefinition: protocolDefinition,
+//     collector: collector,
+//     codeGenerator: (spec) => generateCode(spec, true),
+//     config: config,
+//   );
 
-  bool protocolIsDirty = false;
-  bool libIsDirty = false;
+//   // Analyze the code
+//   var errors = await analyzer.getErrors();
+//   if (errors.isNotEmpty) {
+//     for (var error in errors) {
+//       print(error);
+//     }
+//     return;
+//   }
 
-  var watcher = SourceFileWatcher(
-    onChangedSourceFile: (changedPath, isProtocol) async {
-      protocolIsDirty = protocolIsDirty || isProtocol;
-      libIsDirty = true;
+//   // Generate continuously and hot reload.
+//   var serverRunner = _ServerRunner();
+//   // var dockerRunner = _DockerRunner();
+//   var generatingAndReloading = false;
 
-      // Batch process changes made within 500 ms.
-      while (libIsDirty) {
-        // The timer is used to group together changes if many files are saved
-        // at the same time.
-        Timer(const Duration(milliseconds: 500), () async {
-          if (libIsDirty && !generatingAndReloading) {
-            var protocolWasDirty = protocolIsDirty;
+//   bool protocolIsDirty = false;
+//   bool libIsDirty = false;
 
-            generatingAndReloading = true;
-            protocolIsDirty = false;
-            libIsDirty = false;
+//   var watcher = SourceFileWatcher(
+//     onChangedSourceFile: (changedPath, isProtocol) async {
+//       protocolIsDirty = protocolIsDirty || isProtocol;
+//       libIsDirty = true;
 
-            await _generateAndReload(
-              verbose,
-              protocolWasDirty,
-              configInfo,
-              serverRunner,
-              config: config,
-              analyzer: analyzer,
-            );
+//       // Batch process changes made within 500 ms.
+//       while (libIsDirty) {
+//         // The timer is used to group together changes if many files are saved
+//         // at the same time.
+//         Timer(const Duration(milliseconds: 500), () async {
+//           if (libIsDirty && !generatingAndReloading) {
+//             var protocolWasDirty = protocolIsDirty;
 
-            generatingAndReloading = false;
-          }
-        });
-        await Future.delayed(const Duration(seconds: 1));
-      }
-    },
-    onRemovedProtocolFile: (removedPath) async {
-      // TODO: remove corresponding file
-    },
-    config: config,
-  );
+//             generatingAndReloading = true;
+//             protocolIsDirty = false;
+//             libIsDirty = false;
 
-  // Start Docker.
-  // if (runDocker) {
-  //   print('Starting Docker (for Postgres and Redis).');
-  //   await dockerRunner.start(verbose);
-  // }
+//             await _generateAndReload(
+//               verbose,
+//               protocolWasDirty,
+//               configInfo,
+//               serverRunner,
+//               config: config,
+//               analyzer: analyzer,
+//             );
 
-  // Verify that Postgres & Redis is up and running.
-  print(
-      'Waiting for Postgres on ${configInfo.config.database.host}:${configInfo.config.database.port}.');
-  if (!await PortScanner.waitForPort(
-    configInfo.config.database.host,
-    configInfo.config.database.port,
-    printProgress: true,
-  )) {
-    print('Failed to connect to Postgres.');
-    return;
-  }
+//             generatingAndReloading = false;
+//           }
+//         });
+//         await Future.delayed(const Duration(seconds: 1));
+//       }
+//     },
+//     onRemovedProtocolFile: (removedPath) async {
+//       // TODO: remove corresponding file
+//     },
+//     config: config,
+//   );
 
-  print(
-      'Waiting for Redis on ${configInfo.config.redis.host}:${configInfo.config.redis.port}.');
-  if (!await PortScanner.waitForPort(
-    configInfo.config.redis.host,
-    configInfo.config.redis.port,
-    printProgress: true,
-  )) {
-    print('Failed to connect to Redis.');
-    return;
-  }
+//   // Start Docker.
+//   // if (runDocker) {
+//   //   print('Starting Docker (for Postgres and Redis).');
+//   //   await dockerRunner.start(verbose);
+//   // }
 
-  // Start the server.
-  print('Setup complete. Starting the server.');
-  print('');
-  await serverRunner.start();
+//   // Verify that Postgres & Redis is up and running.
+//   print(
+//       'Waiting for Postgres on ${configInfo.config.database.host}:${configInfo.config.database.port}.');
+//   if (!await PortScanner.waitForPort(
+//     configInfo.config.database.host,
+//     configInfo.config.database.port,
+//     printProgress: true,
+//   )) {
+//     print('Failed to connect to Postgres.');
+//     return;
+//   }
 
-  // Start watching the source directories.
-  unawaited(watcher.watch(verbose));
-}
+//   print(
+//       'Waiting for Redis on ${configInfo.config.redis.host}:${configInfo.config.redis.port}.');
+//   if (!await PortScanner.waitForPort(
+//     configInfo.config.redis.host,
+//     configInfo.config.redis.port,
+//     printProgress: true,
+//   )) {
+//     print('Failed to connect to Redis.');
+//     return;
+//   }
 
-Future<void> _generateAndReload(
-  bool verbose,
-  bool generate,
-  ConfigInfo configInfo,
-  _ServerRunner runner, {
-  required GeneratorConfig config,
-  required ProtocolDartFileAnalyzer analyzer,
-}) async {
-  print('');
-  print('Attempting to generate code and restart server. Hang tight.');
+//   // Start the server.
+//   print('Setup complete. Starting the server.');
+//   print('');
+//   await serverRunner.start();
 
-  if (generate) {
-    try {
-      // TODO: Fix!
-      // performGenerateClasses(verbose);
-    } catch (e, stackTrace) {
-      print('Failed to generate classes: $e');
-      print(stackTrace);
-    }
+//   // Start watching the source directories.
+//   unawaited(watcher.watch(verbose));
+// }
 
-    try {
-      var collector = CodeAnalysisCollector();
-      var protocolDefinition = await analyzer.analyze(
-        verbose: verbose,
-        collector: collector,
-        changedFiles: {},
-      );
+// Future<void> _generateAndReload(
+//   bool verbose,
+//   bool generate,
+//   ConfigInfo configInfo,
+//   _ServerRunner runner, {
+//   required GeneratorConfig config,
+//   required EndpointsAnalyzer analyzer,
+// }) async {
+//   print('');
+//   print('Attempting to generate code and restart server. Hang tight.');
 
-      collector.printErrors();
-      await performGenerateProtocol(
-        verbose: verbose,
-        protocolDefinition: protocolDefinition,
-        collector: collector,
-        codeGenerator: (spec) => generateCode(spec, true),
-        config: config,
-      );
-    } catch (e, stackTrace) {
-      print('Failed to generate protocol: $e');
-      print(stackTrace);
-    }
-  }
+//   if (generate) {
+//     try {
+//       // TODO: Fix!
+//       // performGenerateClasses(verbose);
+//     } catch (e, stackTrace) {
+//       print('Failed to generate classes: $e');
+//       print(stackTrace);
+//     }
 
-  try {
-    // TODO: Implement real hot reload
-    // Check if server code is valid before restarting.
-    var errors = await analyzer.getErrors();
+//     try {
+//       var collector = CodeGenerationCollector();
+//       var protocolDefinition = await analyzer.analyze(
+//         verbose: verbose,
+//         collector: collector,
+//         changedFiles: {},
+//       );
 
-    if (errors.isNotEmpty) {
-      print('Server restart failed.');
-      for (var error in errors) {
-        print(error);
-      }
-    } else {
-      print('Stopping the server.');
-      await runner.stop();
-      await Future.delayed(const Duration(milliseconds: 500));
+//       collector.printErrors();
+//       await performGenerateProtocol(
+//         verbose: verbose,
+//         protocolDefinition: protocolDefinition,
+//         collector: collector,
+//         codeGenerator: (spec) => generateCode(spec, true),
+//         config: config,
+//       );
+//     } catch (e, stackTrace) {
+//       print('Failed to generate protocol: $e');
+//       print(stackTrace);
+//     }
+//   }
 
-      // Start a new instance of the server
-      print('Restarting the server.');
-      print('');
-      await runner.start();
-    }
-  } catch (e) {
-    print('Failed hot reload: $e');
-  }
-}
+//   try {
+//     // TODO: Implement real hot reload
+//     // Check if server code is valid before restarting.
+//     var errors = await analyzer.getErrors();
 
-class _ServerRunner {
-  Process? _process;
+//     if (errors.isNotEmpty) {
+//       print('Server restart failed.');
+//       for (var error in errors) {
+//         print(error);
+//       }
+//     } else {
+//       print('Stopping the server.');
+//       await runner.stop();
+//       await Future.delayed(const Duration(milliseconds: 500));
 
-  Future<void> start() async {
-    assert(_process == null);
-    _process = await Process.start(
-      'dart',
-      ['bin/main.dart'],
-      mode: ProcessStartMode.inheritStdio,
-      runInShell: true,
-    );
-  }
+//       // Start a new instance of the server
+//       print('Restarting the server.');
+//       print('');
+//       await runner.start();
+//     }
+//   } catch (e) {
+//     print('Failed hot reload: $e');
+//   }
+// }
 
-  Future<void> stop() async {
-    // TODO: First attempt to use the shutdown command (needs to be fixed).
+// class _ServerRunner {
+//   Process? _process;
 
-    assert(_process != null);
-    await _process!.killAll();
-    await _process!.exitCode;
-    _process = null;
-  }
+//   Future<void> start() async {
+//     assert(_process == null);
+//     _process = await Process.start(
+//       'dart',
+//       ['bin/main.dart'],
+//       mode: ProcessStartMode.inheritStdio,
+//       runInShell: true,
+//     );
+//   }
 
-  Future<int> exitCode() async {
-    return _process!.exitCode;
-  }
-}
+//   Future<void> stop() async {
+//     // TODO: First attempt to use the shutdown command (needs to be fixed).
+
+//     assert(_process != null);
+//     await _process!.killAll();
+//     await _process!.exitCode;
+//     _process = null;
+//   }
+
+//   Future<int> exitCode() async {
+//     return _process!.exitCode;
+//   }
+// }
